more updates of verletSkin variable to verletSkinPerTimestep

+++ src/autopas/containers/directSum/DirectSum.h
@@ -53,7 +53,7 @@ class DirectSum : public CellBasedParticleContainer<FullParticleCell<Particle>> 
    * @param cutoff
    * @param skinpPerTimestep
    */
-  DirectSum(const std::array<double, 3> boxMin, const std::array<double, 3> boxMax, double cutoff, double skinPerTimestep, double vertletRebuildFrequency)
+  DirectSum(const std::array<double, 3> boxMin, const std::array<double, 3> boxMax, double cutoff, double skinPerTimestep, unsigned int vertletRebuildFrequency)
       : CellBasedParticleContainer<ParticleCell>(boxMin, boxMax, cutoff, skinPerTimestep, vertletRebuildFrequency), _cellBorderFlagManager() {
     this->_cells.resize(2);
   }
